# [rospy] Improve rospy.logXXX_throttle performance