@dirk-thomas 2017-07-07
Thank you for the improvement.

@wkentaro Since you implemented this feature could you give this patch a try since the tests don't cover the actual frame information.