@wkentaro 2017-07-09
LGTM. It works and improves the performance.