@otamachan 2017-07-10
@wkentaro Thanks for taking your time and testing.